move layersData in separate file
